refactor: move linter checkers to a dedicated package
`robocop.linter.rules` used to hold both rule definitions and the AST
checkers that visit the code. After the visitor merging waves, some
modules were purely declarative, some purely visitors and some mixed,
which made the layout hard to follow.

All 23 checkers now live in the new `robocop.linter.checkers` package,
with one module per rule category mirroring `robocop.linter.rules`.
`robocop.linter.rules` is now purely declarative: rules, rule parameters
and small helpers. The dependency direction is strictly
checkers -> rules, so there are no cycles.

`robocop.linter.rules` remains the public import path for the base
classes (`Rule`, `VisitorChecker`, ...) used by custom rules, so this
is not a breaking change for users.

`LinterImporter` now imports both packages: rule modules first (so the
`robocop.linter.rules` package attributes are rebound to the fresh
modules), then the checker modules. Both prefixes are purged from
`sys.modules` before each run to keep the existing guarantee of fresh
`Rule` class objects per run. `DocumentationImporter` only walks the
rule modules.

Linter output is unchanged: verified byte-identical for a full
`--select ALL` scan including disabled-by-default rules, and for
`--fix --diff`.
