+++ src/robocop/linter/checkers/__init__.py
@@ -1,0 +1,1 @@
+"""Checkers walk the parsed Robot Framework files and call the rules defined in ``robocop.linter.rules``."""
