+++ src/robocop/linter/checkers/names.py
