+++ src/robocop/linter/checkers/raw_file.py
