+++ src/robocop/linter/checkers/sections.py
