+++ src/robocop/linter/checkers/settings.py
