+++ src/robocop/linter/checkers/variable_statements.py
